split syntax task and run no-es6-dist on true publish files
